env as integer

+++ ScenarioSetup_SWO2025.R
@@ -54,8 +54,8 @@ source("R/helpers.R")
 ##### Path Setup and check base files ######
 ############################################
 
-batch_count <- as.numeric(Sys.getenv("BATCH_COUNT", 1))
-batch_index <- as.numeric(Sys.getenv("BATCH_INDEX", 1))
+batch_count <- as.integer(Sys.getenv("BATCH_COUNT", 1))
+batch_index <- as.integer(Sys.getenv("BATCH_INDEX", 1))
 
 
 BaseDir <- "diagnostic/base/"
